split a bit
